+++ trie-db/src/query_plan/mod.rs
@@ -24,6 +24,7 @@
 use core::marker::PhantomData;
 
 use crate::{
+	content_proof::Op,
 	nibble::{nibble_ops, nibble_ops::NIBBLE_LENGTH, LeftNibbleSlice, NibbleSlice},
 	node::{NodeHandle, NodePlan, OwnedNode, Value},
 	node_codec::NodeCodec,
@@ -36,12 +37,16 @@ use crate::{
 	},
 	CError, ChildReference, DBValue, NibbleVec, Trie, TrieDB, TrieHash, TrieLayout,
 };
-use crate::content_proof::Op;
 use hash_db::Hasher;
 pub use record::{record_query_plan, HaltedStateRecord, Recorder};
+pub use verify::verify_query_plan_iter;
+use verify::HaltedStateCheckNode;
+pub use verify_content::verify_query_plan_iter_content;
+use verify_content::HaltedStateCheckContent;
 
-pub mod verify_content;
 mod record;
+mod verify;
+mod verify_content;
 
 /// Item to query, in memory.
 #[derive(Default)]
@@ -171,7 +176,7 @@ pub enum ProofKind {
 	*/
 	/// Content oriented proof, no nodes are written, just a
 	/// sequence of accessed by lexicographical order as described
-	/// in verify_content::Op.
+	/// in content_proof::Op.
 	/// As with compact node, checking validity of proof need to
 	/// load the full proof or up to the halt point.
 	CompactContent,
@@ -375,122 +380,12 @@ pub enum HaltedStateCheck<'a, L: TrieLayout, C, D: SplitFirst> {
 	Content(HaltedStateCheckContent<'a, L, C>),
 }
 
-/// When process is halted keep execution state
-/// to restore later.
-pub struct HaltedStateCheckNode<'a, L: TrieLayout, C, D: SplitFirst> {
-	query_plan: QueryPlan<'a, C>,
-	current: Option<QueryPlanItem<'a>>,
-	stack: ReadStack<L, D>,
-	state: ReadProofState,
-	restore_offset: usize,
-}
-
-impl<'a, L: TrieLayout, C, D: SplitFirst> From<QueryPlan<'a, C>>
-	for HaltedStateCheckNode<'a, L, C, D>
-{
-	fn from(query_plan: QueryPlan<'a, C>) -> Self {
-		let is_compact = match query_plan.kind {
-			ProofKind::FullNodes => false,
-			ProofKind::CompactNodes => true,
-			_ => false,
-		};
-
-		HaltedStateCheckNode {
-			stack: ReadStack {
-				items: Default::default(),
-				start_items: 0,
-				prefix: Default::default(),
-				is_compact,
-				expect_value: false,
-				iter_prefix: None,
-				_ph: PhantomData,
-			},
-			state: ReadProofState::NotStarted,
-			current: None,
-			restore_offset: 0,
-			query_plan,
-		}
-	}
-}
-
-/// When process is halted keep execution state
-/// to restore later.
-pub struct HaltedStateCheckContent<'a, L: TrieLayout, C> {
-	query_plan: QueryPlan<'a, C>,
-	current: Option<QueryPlanItem<'a>>,
-	stack: ReadContentStack<L>,
-	state: ReadProofState,
-	restore_offset: usize,
-}
-
-impl<'a, L: TrieLayout, C> From<QueryPlan<'a, C>> for HaltedStateCheckContent<'a, L, C> {
-	fn from(query_plan: QueryPlan<'a, C>) -> Self {
-		HaltedStateCheckContent {
-			stack: ReadContentStack {
-				items: Default::default(),
-				start_items: 0,
-				prefix: Default::default(),
-				expect_value: false,
-				iter_prefix: None,
-				is_prev_push_key: false,
-				is_prev_pop_key: false,
-				is_prev_hash_child: None,
-				first: true,
-				_ph: PhantomData,
-			},
-			state: ReadProofState::NotStarted,
-			current: None,
-			restore_offset: 0,
-			query_plan,
-		}
-	}
-}
-
 enum TryStackChildResult {
 	Stacked,
 	NotStackedBranch,
 	NotStacked,
 	StackedDescendIncomplete,
 	Halted,
-}
-
-/// Proof reading iterator.
-pub struct ReadProofIterator<'a, L, C, D, P>
-where
-	L: TrieLayout,
-	C: Iterator<Item = QueryPlanItem<'a>>,
-	P: Iterator<Item = D>,
-	D: SplitFirst,
-{
-	// always needed, this is option only
-	// to avoid unsafe code when halting.
-	query_plan: Option<QueryPlan<'a, C>>,
-	proof: P,
-	is_compact: bool,
-	expected_root: Option<TrieHash<L>>,
-	current: Option<QueryPlanItem<'a>>,
-	state: ReadProofState,
-	stack: ReadStack<L, D>,
-	restore_offset: usize,
-}
-
-/// Proof reading iterator.
-pub struct ReadProofContentIterator<'a, L, C, P>
-where
-	L: TrieLayout,
-	C: Iterator<Item = QueryPlanItem<'a>>,
-	P: Iterator<Item = Option<Op<TrieHash<L>, Vec<u8>>>>,
-{
-	// always needed, this is option only
-	// to avoid unsafe code when halting.
-	query_plan: Option<QueryPlan<'a, C>>,
-	proof: P,
-	expected_root: Option<TrieHash<L>>,
-	current: Option<QueryPlanItem<'a>>,
-	state: ReadProofState,
-	stack: ReadContentStack<L>,
-	restore_offset: usize,
-	buf_op: Option<Op<TrieHash<L>, Vec<u8>>>,
 }
 
 #[derive(Eq, PartialEq)]
@@ -1339,7 +1234,6 @@ impl<L: TrieLayout> ReadContentStack<L> {
 		let ItemContentStack { children, value: v, depth, .. } = self.items.pop().expect("checked");
 
 		debug_assert!(branch_d == depth);
-		let pr = NibbleSlice::new_offset(&key_branch, branch_d);
 
 		let hashed;
 		let value = if let Some(v) = v.as_ref() {
@@ -1490,666 +1384,4 @@ pub enum ReadProofItem<'a, L: TrieLayout, C, D: SplitFirst> {
 	/// End of a previously start prefix.
 	/// TODO unused implement
 	EndPrefix,
-}
-
-impl<'a, L, C, D, P> Iterator for ReadProofIterator<'a, L, C, D, P>
-where
-	L: TrieLayout,
-	C: Iterator<Item = QueryPlanItem<'a>>,
-	P: Iterator<Item = D>,
-	D: SplitFirst,
-{
-	type Item = Result<ReadProofItem<'a, L, C, D>, VerifyError<TrieHash<L>, CError<L>>>;
-
-	fn next(&mut self) -> Option<Self::Item> {
-		if self.state == ReadProofState::Finished {
-			return None
-		}
-		let check_hash = self.expected_root.is_some();
-		if self.state == ReadProofState::Halted {
-			self.state = ReadProofState::Running;
-		}
-		let mut to_check_slice = self
-			.current
-			.as_ref()
-			.map(|n| NibbleSlice::new_offset(n.key, self.restore_offset));
-
-		// read proof
-		loop {
-			if self.state == ReadProofState::SwitchQueryPlan ||
-				self.state == ReadProofState::NotStarted
-			{
-				let query_plan = self.query_plan.as_mut().expect("Removed with state");
-				if let Some(next) = query_plan.items.next() {
-					let (ordered, common_nibbles) = if let Some(old) = self.current.as_ref() {
-						old.before(&next)
-					} else {
-						(true, 0)
-					};
-					if !ordered {
-						if query_plan.ignore_unordered {
-							continue
-						} else {
-							self.state = ReadProofState::Finished;
-							return Some(Err(VerifyError::UnorderedKey(next.key.to_vec())))
-						}
-					}
-
-					let r = self.stack.pop_until(common_nibbles, &self.expected_root, false);
-					if let Err(e) = r {
-						self.state = ReadProofState::Finished;
-						return Some(Err(e))
-					}
-					self.state = ReadProofState::Running;
-					self.current = Some(next);
-					to_check_slice = self
-						.current
-						.as_ref()
-						.map(|n| NibbleSlice::new_offset(n.key, common_nibbles));
-				} else {
-					self.state = ReadProofState::PlanConsumed;
-					self.current = None;
-					break
-				}
-			};
-			let did_prefix = self.stack.iter_prefix.is_some();
-			while let Some((_, accessed_value_node, hash_only)) = self.stack.iter_prefix.clone() {
-				// prefix iteration
-				if !accessed_value_node {
-					self.stack.iter_prefix.as_mut().map(|s| {
-						s.1 = true;
-					});
-					match self.stack.access_value(&mut self.proof, check_hash, hash_only) {
-						Ok((Some(value), None)) =>
-							return Some(Ok(ReadProofItem::Value(
-								self.stack.prefix.inner().to_vec().into(),
-								value,
-							))),
-						Ok((None, Some(hash))) =>
-							return Some(Ok(ReadProofItem::Hash(
-								self.stack.prefix.inner().to_vec().into(),
-								hash,
-							))),
-						Ok((None, None)) => (),
-						Ok(_) => unreachable!(),
-						Err(e) => {
-							self.state = ReadProofState::Finished;
-							return Some(Err(e))
-						},
-					};
-				}
-				while let Some(child_index) = self.stack.items.last_mut().and_then(|last| {
-					if last.next_descended_child as usize >= NIBBLE_LENGTH {
-						None
-					} else {
-						let child_index = last.next_descended_child;
-						last.next_descended_child += 1;
-						Some(child_index)
-					}
-				}) {
-					let r = match self.stack.try_stack_child(
-						child_index,
-						&mut self.proof,
-						&self.expected_root,
-						None,
-						false,
-					) {
-						Ok(r) => r,
-						Err(e) => {
-							self.state = ReadProofState::Finished;
-							return Some(Err(e))
-						},
-					};
-					match r {
-						TryStackChildResult::Stacked => {
-							self.stack.iter_prefix.as_mut().map(|p| {
-								p.1 = false;
-							});
-							break
-						},
-						TryStackChildResult::StackedDescendIncomplete => {
-							unreachable!("slice query none");
-						},
-						TryStackChildResult::NotStacked => break,
-						TryStackChildResult::NotStackedBranch => (),
-						TryStackChildResult::Halted => {
-							if let Some(last) = self.stack.items.last_mut() {
-								last.next_descended_child -= 1;
-							}
-							return self.halt(None)
-						},
-					}
-				}
-				if self.stack.iter_prefix.as_ref().map(|p| p.1).unwrap_or_default() {
-					if !match self.stack.pop(&self.expected_root) {
-						Ok(r) => r,
-						Err(e) => {
-							self.state = ReadProofState::Finished;
-							return Some(Err(e))
-						},
-					} {
-						// end iter
-						self.stack.exit_prefix_iter();
-					}
-				}
-			}
-			if did_prefix {
-				// exit a prefix iter, next content looping
-				self.state = ReadProofState::SwitchQueryPlan;
-				continue
-			}
-			let to_check = self.current.as_ref().expect("Init above");
-			let to_check_len = to_check.key.len() * nibble_ops::NIBBLE_PER_BYTE;
-			let mut to_check_slice = to_check_slice.as_mut().expect("Init above");
-			let as_prefix = to_check.as_prefix; // TODO useless?
-			let hash_only = to_check.hash_only; // TODO useless?
-			let mut at_value = false;
-			match self.stack.prefix.len().cmp(&to_check_len) {
-				Ordering::Equal =>
-					if !self.stack.items.is_empty() {
-						at_value = true;
-					},
-				Ordering::Less => (),
-				Ordering::Greater => {
-					unreachable!();
-				},
-			}
-
-			if at_value {
-				if as_prefix {
-					self.stack.enter_prefix_iter(hash_only);
-					continue
-				}
-				self.state = ReadProofState::SwitchQueryPlan;
-				match self.stack.access_value(&mut self.proof, check_hash, hash_only) {
-					Ok((Some(value), None)) =>
-						return Some(Ok(ReadProofItem::Value(to_check.key.into(), value))),
-					Ok((None, Some(hash))) =>
-						return Some(Ok(ReadProofItem::Hash(to_check.key.into(), hash))),
-					Ok((None, None)) => return Some(Ok(ReadProofItem::NoValue(to_check.key))),
-					Ok(_) => unreachable!(),
-					Err(e) => {
-						self.state = ReadProofState::Finished;
-						return Some(Err(e))
-					},
-				}
-			}
-
-			let child_index = if self.stack.items.len() == 0 {
-				// dummy
-				0
-			} else {
-				to_check_slice.at(0)
-			};
-			let r = match self.stack.try_stack_child(
-				child_index,
-				&mut self.proof,
-				&self.expected_root,
-				Some(&mut to_check_slice),
-				to_check.as_prefix,
-			) {
-				Ok(r) => r,
-				Err(e) => {
-					self.state = ReadProofState::Finished;
-					return Some(Err(e))
-				},
-			};
-			match r {
-				TryStackChildResult::Stacked => (),
-				TryStackChildResult::StackedDescendIncomplete => {
-					if as_prefix {
-						self.stack.enter_prefix_iter(hash_only);
-						continue
-					}
-					self.state = ReadProofState::SwitchQueryPlan;
-					return Some(Ok(ReadProofItem::NoValue(to_check.key)))
-				},
-				TryStackChildResult::NotStacked => {
-					self.state = ReadProofState::SwitchQueryPlan;
-					return Some(Ok(ReadProofItem::NoValue(to_check.key)))
-				},
-				TryStackChildResult::NotStackedBranch => {
-					self.state = ReadProofState::SwitchQueryPlan;
-					return Some(Ok(ReadProofItem::NoValue(to_check.key)))
-				},
-				TryStackChildResult::Halted => return self.halt(Some(to_check_slice)),
-			}
-		}
-
-		debug_assert!(self.state == ReadProofState::PlanConsumed);
-		if self.is_compact {
-			let stack_to = 0; // TODO restart is different
-				  //					let r = self.stack.pop_until(common_nibbles, &self.expected_root);
-			let r = self.stack.pop_until(stack_to, &self.expected_root, false);
-			if let Err(e) = r {
-				self.state = ReadProofState::Finished;
-				return Some(Err(e))
-			}
-		} else {
-			if self.proof.next().is_some() {
-				self.state = ReadProofState::Finished;
-				return Some(Err(VerifyError::ExtraneousNode))
-			}
-		}
-		self.state = ReadProofState::Finished;
-		return None
-	}
-}
-
-impl<'a, L, C, P> Iterator for ReadProofContentIterator<'a, L, C, P>
-where
-	L: TrieLayout,
-	C: Iterator<Item = QueryPlanItem<'a>>,
-	P: Iterator<Item = Option<Op<TrieHash<L>, Vec<u8>>>>,
-{
-	type Item = Result<ReadProofItem<'a, L, C, Vec<u8>>, VerifyError<TrieHash<L>, CError<L>>>;
-
-	fn next(&mut self) -> Option<Self::Item> {
-		if self.state == ReadProofState::Finished {
-			return None
-		}
-		let check_hash = self.expected_root.is_some();
-		if self.state == ReadProofState::Halted {
-			self.state = ReadProofState::Running;
-		}
-		let mut to_check_slice = self
-			.current
-			.as_ref()
-			.map(|n| NibbleSlice::new_offset(n.key, self.restore_offset));
-		// read proof
-		loop {
-			if self.state == ReadProofState::SwitchQueryPlan ||
-				self.state == ReadProofState::NotStarted
-			{
-				let query_plan = self.query_plan.as_mut().expect("Removed with state");
-				if let Some(next) = query_plan.items.next() {
-					let (ordered, common_nibbles) = if let Some(old) = self.current.as_ref() {
-						old.before(&next)
-					} else {
-						(true, 0)
-					};
-					if !ordered {
-						if query_plan.ignore_unordered {
-							continue
-						} else {
-							self.state = ReadProofState::Finished;
-							return Some(Err(VerifyError::UnorderedKey(next.key.to_vec())))
-						}
-					}
-
-					let r = self.stack.pop_until(common_nibbles, false);
-					if let Err(e) = r {
-						self.state = ReadProofState::Finished;
-						return Some(Err(e))
-					}
-					self.state = ReadProofState::Running;
-					self.current = Some(next);
-					to_check_slice = self
-						.current
-						.as_ref()
-						.map(|n| NibbleSlice::new_offset(n.key, common_nibbles));
-				} else {
-					self.state = ReadProofState::PlanConsumed;
-					self.current = None;
-				}
-			};
-
-			// op sequence check
-			let mut from_iter = false;
-			while let Some(op) = self.buf_op.take().map(Option::Some).or_else(|| {
-				from_iter = true;
-				self.proof.next()
-			}) {
-				println!("read: {:?}", op);
-				let Some(op) = op else {
-					let r = self.stack.stack_pop(None, &self.expected_root);
-						self.state = ReadProofState::Finished;
-						if let Err(e) = r {
-							self.state = ReadProofState::Finished;
-							return Some(Err(e))
-						}
-						if let Some(c) = self.current.as_ref() {
-						if c.as_prefix {
-							// end prefix switch to next
-							self.state = ReadProofState::SwitchQueryPlan;
-							break;
-						} else {
-							// missing value
-							self.state = ReadProofState::SwitchQueryPlan;
-							return Some(Ok(ReadProofItem::NoValue(&c.key)));
-						}
-					} else {
-						return None; // finished
-					}
-				};
-				if from_iter {
-					// check ordering logic
-					// TODO wrap in an error and put bools in a struct
-					match &op {
-						Op::KeyPush(..) => {
-							if self.stack.is_prev_push_key {
-								self.state = ReadProofState::Finished;
-								return Some(Err(VerifyError::ExtraneousNode)) // TODO a decode op error
-								              // TODO return
-								              // Err(CompactDecoderError::ConsecutivePushKeys.
-								              // into())
-							}
-							self.stack.is_prev_push_key = true;
-							self.stack.is_prev_pop_key = false;
-							self.stack.is_prev_hash_child = None;
-							self.stack.first = false;
-						},
-						Op::KeyPop(..) => {
-							if self.stack.is_prev_pop_key {
-								self.state = ReadProofState::Finished;
-								return Some(Err(VerifyError::ExtraneousNode)) // TODO a decode op error
-								              // return Err(CompactDecoderError::ConsecutivePopKeys.
-								              // into())
-							}
-							self.stack.is_prev_push_key = false;
-							self.stack.is_prev_pop_key = true;
-							self.stack.is_prev_hash_child = None;
-							self.stack.first = false;
-						},
-						Op::HashChild(_, ix) => {
-							if let Some(prev_ix) = self.stack.is_prev_hash_child.as_ref() {
-								if prev_ix >= ix {
-									self.state = ReadProofState::Finished;
-									return Some(Err(VerifyError::ExtraneousNode)) // TODO a decode op error
-									          // return Err(CompactDecoderError::NotConsecutiveHash.
-									          // into())
-								}
-							}
-							// child ix on an existing content would be handle by iter_build.
-							self.stack.is_prev_push_key = false;
-							self.stack.is_prev_pop_key = false;
-							self.stack.is_prev_hash_child = Some(*ix);
-						},
-						Op::Value(_) => {
-							//	| Op::ValueForceInline(_) | Op::ValueForceHashed(_) => {
-							if !(self.stack.is_prev_push_key || self.stack.first) {
-								self.state = ReadProofState::Finished;
-								return Some(Err(VerifyError::ExtraneousNode)) // TODO a decode op error
-								              // return Err(CompactDecoderError::ValueNotAfterPush.
-								              // into())
-							}
-							self.stack.is_prev_push_key = false;
-							self.stack.is_prev_pop_key = false;
-							self.stack.is_prev_hash_child = None;
-							self.stack.first = false;
-						},
-						_ => {
-							self.stack.is_prev_push_key = false;
-							self.stack.is_prev_pop_key = false;
-							self.stack.is_prev_hash_child = None;
-							self.stack.first = false;
-						},
-					}
-
-					// debug TODO make it log and external function
-					match &op {
-						Op::HashChild(hash, child_ix) => {
-							println!(
-								"ChildHash {:?}, {:?}, {:?}",
-								self.stack.prefix, child_ix, hash
-							);
-						},
-						Op::HashValue(hash) => {
-							println!("ValueHash {:?}, {:?}", self.stack.prefix, hash);
-						},
-						Op::Value(value) => {
-							println!("Value {:?}, {:?}", self.stack.prefix, value);
-						},
-						_ => (),
-					}
-				}
-				from_iter = false;
-
-				// next
-				let item = if match &op {
-					Op::Value(..) | Op::HashValue(..) => true,
-					_ => false,
-				} {
-					let mut to_check_slice = to_check_slice.as_mut().expect("Init above");
-
-					let mut at_value = false;
-					let mut next_query = false;
-					if let Some(current) = self.current.as_ref() {
-						let query_slice = LeftNibbleSlice::new(&current.key);
-						match self.stack.prefix.as_leftnibbleslice().cmp(&query_slice) {
-							Ordering::Equal =>
-								if !self.stack.items.is_empty() {
-									at_value = true;
-								},
-							Ordering::Less => (),
-							Ordering::Greater =>
-								if current.as_prefix {
-									let query_slice = LeftNibbleSlice::new(&current.key);
-									if self
-										.stack
-										.prefix
-										.as_leftnibbleslice()
-										.starts_with(&query_slice)
-									{
-										at_value = true;
-									} else {
-										next_query = true;
-									}
-								} else {
-									next_query = true;
-								},
-						}
-						if next_query {
-							self.buf_op = Some(op);
-							self.state = ReadProofState::SwitchQueryPlan;
-							if current.as_prefix {
-								break
-							} else {
-								return Some(Ok(ReadProofItem::NoValue(&current.key)))
-							}
-						}
-					}
-
-					if at_value {
-						match &op {
-							Op::Value(value) => {
-								// TODO could get content from op with no clone.
-								Some(ReadProofItem::Value(
-									self.stack.prefix.inner().to_vec().into(),
-									value.clone(),
-								))
-							},
-							Op::HashValue(hash) => {
-								// TODO could get content from op with no clone.
-								Some(ReadProofItem::Hash(
-									self.stack.prefix.inner().to_vec().into(),
-									hash.clone(),
-								))
-							},
-							_ => unreachable!(),
-						}
-					} else {
-						match &op {
-							Op::Value(value) => {
-								// hash value here not value
-								if L::MAX_INLINE_VALUE
-									.map(|max| max as usize <= value.len())
-									.unwrap_or(false)
-								{
-									self.state = ReadProofState::Finished;
-									return Some(Err(VerifyError::ExtraneousValue(value.clone())))
-								}
-							},
-							_ => (),
-						}
-						None
-					}
-				} else {
-					None
-				};
-
-				// act
-				let r = match op {
-					Op::KeyPush(partial, mask) => {
-						self.stack
-							.prefix
-							.append_slice(LeftNibbleSlice::new_with_mask(partial.as_slice(), mask));
-						self.stack.stack_empty(self.stack.prefix.len());
-						Ok(())
-					},
-					Op::KeyPop(nb_nibble) => {
-						let r = self.stack.stack_pop(Some(nb_nibble as usize), &self.expected_root);
-						self.stack.prefix.drop_lasts(nb_nibble.into());
-						r
-					},
-					Op::EndProof => break,
-					op => self.stack.set_cache_change(op.into()),
-				};
-				if let Err(e) = r {
-					self.state = ReadProofState::Finished;
-					return Some(Err(e))
-				}
-				if let Some(r) = item {
-					if self.current.as_ref().map(|c| !c.as_prefix).unwrap_or(true) {
-						self.state = ReadProofState::SwitchQueryPlan; // TODO this is same as content NOne?
-					}
-					return Some(Ok(r))
-				}
-			}
-			if self.state != ReadProofState::SwitchQueryPlan && self.current.is_some() {
-				// TODO return halt instead
-				return Some(Err(VerifyError::IncompleteProof))
-			}
-			self.state = ReadProofState::SwitchQueryPlan;
-		}
-
-		self.state = ReadProofState::Finished;
-		if self.proof.next().is_some() {
-			return Some(Err(VerifyError::ExtraneousNode))
-		} else {
-			return None
-		}
-	}
-}
-
-impl<'a, L, C, D, P> ReadProofIterator<'a, L, C, D, P>
-where
-	L: TrieLayout,
-	C: Iterator<Item = QueryPlanItem<'a>>,
-	P: Iterator<Item = D>,
-	D: SplitFirst,
-{
-	fn halt(
-		&mut self,
-		to_check_slice: Option<&mut NibbleSlice>,
-	) -> Option<Result<ReadProofItem<'a, L, C, D>, VerifyError<TrieHash<L>, CError<L>>>> {
-		if self.is_compact {
-			let stack_to = 0; // TODO restart is different
-			let r = self.stack.pop_until(stack_to, &self.expected_root, true);
-			if let Err(e) = r {
-				self.state = ReadProofState::Finished;
-				return Some(Err(e))
-			}
-		}
-		self.state = ReadProofState::Finished;
-		let query_plan = crate::rstd::mem::replace(&mut self.query_plan, None);
-		let query_plan = query_plan.expect("Init with state");
-		let current = crate::rstd::mem::take(&mut self.current);
-		let mut stack = crate::rstd::mem::replace(
-			&mut self.stack,
-			ReadStack {
-				items: Default::default(),
-				start_items: 0,
-				prefix: Default::default(),
-				is_compact: self.is_compact,
-				expect_value: false,
-				iter_prefix: None,
-				_ph: PhantomData,
-			},
-		);
-		stack.start_items = stack.items.len();
-		Some(Ok(ReadProofItem::Halted(Box::new(HaltedStateCheck::Node(HaltedStateCheckNode {
-			query_plan,
-			current,
-			stack,
-			state: ReadProofState::Halted,
-			restore_offset: to_check_slice.map(|s| s.offset()).unwrap_or(0),
-		})))))
-	}
-}
-
-/// Read the proof.
-///
-/// If expected root is None, then we do not check hashes at all.
-pub fn verify_query_plan_iter<'a, L, C, D, P>(
-	state: HaltedStateCheck<'a, L, C, D>,
-	proof: P,
-	expected_root: Option<TrieHash<L>>,
-) -> Result<ReadProofIterator<'a, L, C, D, P>, VerifyError<TrieHash<L>, CError<L>>>
-where
-	L: TrieLayout,
-	C: Iterator<Item = QueryPlanItem<'a>>,
-	P: Iterator<Item = D>,
-	D: SplitFirst,
-{
-	let HaltedStateCheck::Node(state) = state else {
-		return Err(VerifyError::IncompleteProof) // TODO not kind as param if keeping CompactContent
-	};
-	let HaltedStateCheckNode { query_plan, current, stack, state, restore_offset } = state;
-
-	match query_plan.kind {
-		ProofKind::CompactContent => {
-			return Err(VerifyError::IncompleteProof) // TODO not kind as param if keeping CompactContent
-		},
-		_ => (),
-	};
-
-	Ok(ReadProofIterator {
-		query_plan: Some(query_plan),
-		proof,
-		is_compact: stack.is_compact,
-		expected_root,
-		current,
-		state,
-		stack,
-		restore_offset,
-	})
-}
-
-/// Read the proof.
-///
-/// If expected root is None, then we do not check hashes at all.
-pub fn verify_query_plan_iter_content<'a, L, C, P>(
-	state: HaltedStateCheck<'a, L, C, Vec<u8>>,
-	proof: P,
-	expected_root: Option<TrieHash<L>>,
-) -> Result<ReadProofContentIterator<'a, L, C, P>, VerifyError<TrieHash<L>, CError<L>>>
-where
-	L: TrieLayout,
-	C: Iterator<Item = QueryPlanItem<'a>>,
-	P: Iterator<Item = Option<Op<TrieHash<L>, Vec<u8>>>>,
-{
-	let HaltedStateCheck::Content(state) = state else {
-		return Err(VerifyError::IncompleteProof) // TODO not kind as param if keeping CompactContent
-	};
-
-	let HaltedStateCheckContent { query_plan, current, stack, state, restore_offset } = state;
-
-	match query_plan.kind {
-		ProofKind::CompactContent => (),
-		_ => {
-			return Err(VerifyError::IncompleteProof) // TODO not kind as param if keeping CompactContent
-		},
-	};
-
-	Ok(ReadProofContentIterator {
-		query_plan: Some(query_plan),
-		proof,
-		expected_root,
-		current,
-		state,
-		stack,
-		restore_offset,
-		buf_op: None,
-	})
 }
