+++ trie-db/src/query_plan/verify_content.rs
@@ -14,242 +14,384 @@
 
 //! Verify query plan proof for content proofs.
 
-use super::RecorderOutput;
+use super::*;
 use core::marker::PhantomData;
 
-/// Representation of each encoded action
-/// for building the proof.
-/// TODO ref variant for encoding ?? or key using V and use Op<&H, &[u8]>.
-#[derive(Debug)]
-pub enum Op<H, V> {
-	// key content followed by a mask for last byte.
-	// If mask erase some content the content need to
-	// be set at 0 (or error).
-	// Two consecutive `KeyPush` are invalid.
-	KeyPush(Vec<u8>, u8), /* TODO could use BackingByteVec (but Vec for new as it scale
-	                       * encode) */
-	// Last call to pop is implicit (up to root), defining
-	// one will result in an error.
-	// Two consecutive `KeyPop` are invalid.
-	// TODO should be compact encoding of number.
-	KeyPop(u16),
-	// u8 is child index, shorthand for key push one nibble followed by key pop.
-	HashChild(H, u8),
-	// All value variant are only after a `KeyPush` or at first position.
-	HashValue(H),
-	Value(V),
-	// This is not strictly necessary, only if the proof is not sized, otherwhise if we know
-	// the stream will end it can be skipped.
-	EndProof,
+use crate::{
+	content_proof::Op, nibble::LeftNibbleSlice, proof::VerifyError, rstd::result::Result, CError,
+	TrieHash, TrieLayout,
+};
+pub use record::{record_query_plan, HaltedStateRecord, Recorder};
+
+/// Proof reading iterator.
+pub struct ReadProofContentIterator<'a, L, C, P>
+where
+	L: TrieLayout,
+	C: Iterator<Item = QueryPlanItem<'a>>,
+	P: Iterator<Item = Option<Op<TrieHash<L>, Vec<u8>>>>,
+{
+	// always needed, this is option only
+	// to avoid unsafe code when halting.
+	query_plan: Option<QueryPlan<'a, C>>,
+	proof: P,
+	expected_root: Option<TrieHash<L>>,
+	current: Option<QueryPlanItem<'a>>,
+	state: ReadProofState,
+	stack: ReadContentStack<L>,
+	buf_op: Option<Op<TrieHash<L>, Vec<u8>>>,
 }
 
-// Limiting size to u32 (could also just use a terminal character).
-#[derive(Debug, PartialEq, Eq)]
-#[repr(transparent)]
-struct VarInt(u32);
+/// Read the proof.
+///
+/// If expected root is None, then we do not check hashes at all.
+pub fn verify_query_plan_iter_content<'a, L, C, P>(
+	state: HaltedStateCheck<'a, L, C, Vec<u8>>,
+	proof: P,
+	expected_root: Option<TrieHash<L>>,
+) -> Result<ReadProofContentIterator<'a, L, C, P>, VerifyError<TrieHash<L>, CError<L>>>
+where
+	L: TrieLayout,
+	C: Iterator<Item = QueryPlanItem<'a>>,
+	P: Iterator<Item = Option<Op<TrieHash<L>, Vec<u8>>>>,
+{
+	let HaltedStateCheck::Content(state) = state else {
+		return Err(VerifyError::IncompleteProof) // TODO not kind as param if keeping CompactContent
+	};
 
-impl VarInt {
-	fn encoded_len(&self) -> usize {
-		if self.0 == 0 {
-			return 1
-		}
-		let len = 32 - self.0.leading_zeros() as usize;
-		if len % 7 == 0 {
-			len / 7
-		} else {
-			len / 7 + 1
-		}
-		/*
-		match self.0 {
-			l if l < 2 ^ 7 => 1, // leading 0: 25
-			l if l < 2 ^ 14 => 2, // leading 0: 18
+	let HaltedStateCheckContent { query_plan, current, stack, state } = state;
 
-			l if l < 2 ^ 21 => 3, // 11
-			l if l < 2 ^ 28 => 4, // 4
-			_ => 5,
-		}
-		*/
-	}
+	match query_plan.kind {
+		ProofKind::CompactContent => (),
+		_ => {
+			return Err(VerifyError::IncompleteProof) // TODO not kind as param if keeping CompactContent
+		},
+	};
 
-	fn encode_into(&self, out: &mut impl RecorderOutput) {
-		let mut to_encode = self.0;
-		for _ in 0..self.encoded_len() - 1 {
-			out.write_bytes(&[0b1000_0000 | to_encode as u8]);
-			to_encode >>= 7;
-		}
-		out.write_bytes(&[to_encode as u8]);
-	}
-
-	fn decode(encoded: &[u8]) -> Result<(Self, usize), ()> {
-		let mut value = 0u32;
-		for (i, byte) in encoded.iter().enumerate() {
-			let last = byte & 0b1000_0000 == 0;
-			value |= ((byte & 0b0111_1111) as u32) << (i * 7);
-			if last {
-				return Ok((VarInt(value), i + 1))
-			}
-		}
-		Err(())
-	}
+	Ok(ReadProofContentIterator {
+		query_plan: Some(query_plan),
+		proof,
+		expected_root,
+		current,
+		state,
+		stack,
+		buf_op: None,
+	})
 }
 
-#[test]
-fn varint_encode_decode() {
-	let mut buf = super::InMemoryRecorder::default();
-	for i in 0..u16::MAX as u32 + 1 {
-		VarInt(i).encode_into(&mut buf);
-		assert_eq!(buf.buffer.len(), VarInt(i).encoded_len());
-		assert_eq!(Ok((VarInt(i), buf.buffer.len())), VarInt::decode(&buf.buffer));
-		buf.buffer.clear();
-	}
-}
-
-impl<H: AsRef<[u8]>, V: AsRef<[u8]>> Op<H, V> {
-	/// Calculate encoded len.
-	pub fn encoded_len(&self) -> usize {
-		let mut len = 1;
-		match self {
-			Op::KeyPush(key, _mask) => {
-				len += VarInt(key.len() as u32).encoded_len();
-				len += key.len();
-				len += 1;
-			},
-			Op::KeyPop(nb) => {
-				len += VarInt(*nb as u32).encoded_len();
-			},
-			Op::HashChild(hash, _at) => {
-				len += hash.as_ref().len();
-				len += 1;
-			},
-			Op::HashValue(hash) => {
-				len += hash.as_ref().len();
-			},
-			Op::Value(value) => {
-				len += VarInt(value.as_ref().len() as u32).encoded_len();
-				len += value.as_ref().len();
-			},
-			Op::EndProof => (),
-		}
-		len
-	}
-
-	/// Write op.
-	pub fn encode_into(&self, out: &mut impl RecorderOutput) {
-		match self {
-			Op::KeyPush(key, mask) => {
-				out.write_bytes(&[0]);
-				VarInt(key.len() as u32).encode_into(out);
-				out.write_bytes(&key);
-				out.write_bytes(&[*mask]);
-			},
-			Op::KeyPop(nb) => {
-				out.write_bytes(&[1]);
-				VarInt(*nb as u32).encode_into(out);
-			},
-			Op::HashChild(hash, at) => {
-				out.write_bytes(&[2]);
-				out.write_bytes(hash.as_ref());
-				out.write_bytes(&[*at]);
-			},
-			Op::HashValue(hash) => {
-				out.write_bytes(&[3]);
-				out.write_bytes(hash.as_ref());
-			},
-			Op::Value(value) => {
-				out.write_bytes(&[4]);
-				let value = value.as_ref();
-				VarInt(value.len() as u32).encode_into(out);
-				out.write_bytes(&value);
-			},
-			Op::EndProof => {
-				out.write_bytes(&[5]);
-			},
-		}
-	}
-}
-
-impl<H: AsRef<[u8]> + AsMut<[u8]> + Default> Op<H, Vec<u8>> {
-	/// Read an op, return op and number byte read. Or error if invalid encoded.
-	pub fn decode(encoded: &[u8]) -> Result<(Self, usize), ()> {
-		let mut i = 0;
-		if i >= encoded.len() {
-			return Err(())
-		}
-		Ok(match encoded[i] {
-			0 => {
-				let (len, offset) = VarInt::decode(&encoded[i + 1..])?;
-				i += 1 + offset;
-				if i + len.0 as usize >= encoded.len() {
-					return Err(())
-				}
-				let key = &encoded[i..i + len.0 as usize];
-				let mask = encoded[i + len.0 as usize];
-				(Op::KeyPush(key.to_vec(), mask), i + len.0 as usize + 1)
-			},
-			1 => {
-				let (len, offset) = VarInt::decode(&encoded[i + 1..])?;
-				if len.0 > u16::MAX as u32 {
-					return Err(())
-				}
-				(Op::KeyPop(len.0 as u16), i + 1 + offset)
-			},
-			2 => {
-				let mut hash = H::default();
-				let end = i + 1 + hash.as_ref().len();
-				if end >= encoded.len() {
-					return Err(())
-				}
-				hash.as_mut().copy_from_slice(&encoded[i + 1..end]);
-				let mask = encoded[end];
-				(Op::HashChild(hash, mask), end + 1)
-			},
-			3 => {
-				let mut hash = H::default();
-				let end = i + 1 + hash.as_ref().len();
-				if end >= encoded.len() {
-					return Err(())
-				}
-				hash.as_mut().copy_from_slice(&encoded[i + 1..end]);
-				(Op::HashValue(hash), end)
-			},
-			4 => {
-				let (len, offset) = VarInt::decode(&encoded[i + 1..])?;
-				i += 1 + offset;
-				if i + len.0 as usize > encoded.len() {
-					return Err(())
-				}
-				let value = &encoded[i..i + len.0 as usize];
-				(Op::Value(value.to_vec()), i + len.0 as usize)
-			},
-			5 => (Op::EndProof, 1),
-			_ => return Err(()),
-		})
-	}
-}
-
-/// Iterator on op from a in memory encoded proof.
-pub struct IterOpProof<H: AsRef<[u8]> + AsMut<[u8]> + Default, B: AsRef<[u8]>>(
-	B,
-	usize,
-	PhantomData<H>,
-);
-
-impl<H: AsRef<[u8]> + AsMut<[u8]> + Default, B: AsRef<[u8]>> From<B> for IterOpProof<H, B> {
-	fn from(b: B) -> Self {
-		Self(b, 0, PhantomData)
-	}
-}
-
-impl<H: AsRef<[u8]> + AsMut<[u8]> + Default, B: AsRef<[u8]>> Iterator for IterOpProof<H, B> {
-	type Item = Option<Op<H, Vec<u8>>>;
+impl<'a, L, C, P> Iterator for ReadProofContentIterator<'a, L, C, P>
+where
+	L: TrieLayout,
+	C: Iterator<Item = QueryPlanItem<'a>>,
+	P: Iterator<Item = Option<Op<TrieHash<L>, Vec<u8>>>>,
+{
+	type Item = Result<ReadProofItem<'a, L, C, Vec<u8>>, VerifyError<TrieHash<L>, CError<L>>>;
 
 	fn next(&mut self) -> Option<Self::Item> {
-		match Op::decode(&self.0.as_ref()[self.1..]) {
-			Ok((op, len)) => {
-				self.1 += len;
-				Some(Some(op))
+		if self.state == ReadProofState::Finished {
+			return None
+		}
+		if self.state == ReadProofState::Halted {
+			self.state = ReadProofState::Running;
+		}
+		// read proof
+		loop {
+			if self.state == ReadProofState::SwitchQueryPlan ||
+				self.state == ReadProofState::NotStarted
+			{
+				let query_plan = self.query_plan.as_mut().expect("Removed with state");
+				if let Some(next) = query_plan.items.next() {
+					let (ordered, common_nibbles) = if let Some(old) = self.current.as_ref() {
+						old.before(&next)
+					} else {
+						(true, 0)
+					};
+					if !ordered {
+						if query_plan.ignore_unordered {
+							continue
+						} else {
+							self.state = ReadProofState::Finished;
+							return Some(Err(VerifyError::UnorderedKey(next.key.to_vec())))
+						}
+					}
+
+					let r = self.stack.pop_until(common_nibbles, false);
+					if let Err(e) = r {
+						self.state = ReadProofState::Finished;
+						return Some(Err(e))
+					}
+					self.state = ReadProofState::Running;
+					self.current = Some(next);
+				} else {
+					self.state = ReadProofState::PlanConsumed;
+					self.current = None;
+				}
+			};
+
+			// op sequence check
+			let mut from_iter = false;
+			while let Some(op) = self.buf_op.take().map(Option::Some).or_else(|| {
+				from_iter = true;
+				self.proof.next()
+			}) {
+				println!("read: {:?}", op);
+				let Some(op) = op else {
+					let r = self.stack.stack_pop(None, &self.expected_root);
+						self.state = ReadProofState::Finished;
+						if let Err(e) = r {
+							self.state = ReadProofState::Finished;
+							return Some(Err(e))
+						}
+						if let Some(c) = self.current.as_ref() {
+						if c.as_prefix {
+							// end prefix switch to next
+							self.state = ReadProofState::SwitchQueryPlan;
+							break;
+						} else {
+							// missing value
+							self.state = ReadProofState::SwitchQueryPlan;
+							return Some(Ok(ReadProofItem::NoValue(&c.key)));
+						}
+					} else {
+						return None; // finished
+					}
+				};
+				if from_iter {
+					// check ordering logic
+					// TODO wrap in an error and put bools in a struct
+					match &op {
+						Op::KeyPush(..) => {
+							if self.stack.is_prev_push_key {
+								self.state = ReadProofState::Finished;
+								return Some(Err(VerifyError::ExtraneousNode)) // TODO a decode op error
+								              // TODO return
+								              // Err(CompactDecoderError::ConsecutivePushKeys.
+								              // into())
+							}
+							self.stack.is_prev_push_key = true;
+							self.stack.is_prev_pop_key = false;
+							self.stack.is_prev_hash_child = None;
+							self.stack.first = false;
+						},
+						Op::KeyPop(..) => {
+							if self.stack.is_prev_pop_key {
+								self.state = ReadProofState::Finished;
+								return Some(Err(VerifyError::ExtraneousNode)) // TODO a decode op error
+								              // return Err(CompactDecoderError::ConsecutivePopKeys.
+								              // into())
+							}
+							self.stack.is_prev_push_key = false;
+							self.stack.is_prev_pop_key = true;
+							self.stack.is_prev_hash_child = None;
+							self.stack.first = false;
+						},
+						Op::HashChild(_, ix) => {
+							if let Some(prev_ix) = self.stack.is_prev_hash_child.as_ref() {
+								if prev_ix >= ix {
+									self.state = ReadProofState::Finished;
+									return Some(Err(VerifyError::ExtraneousNode)) // TODO a decode op error
+									          // return Err(CompactDecoderError::NotConsecutiveHash.
+									          // into())
+								}
+							}
+							// child ix on an existing content would be handle by iter_build.
+							self.stack.is_prev_push_key = false;
+							self.stack.is_prev_pop_key = false;
+							self.stack.is_prev_hash_child = Some(*ix);
+						},
+						Op::Value(_) => {
+							//	| Op::ValueForceInline(_) | Op::ValueForceHashed(_) => {
+							if !(self.stack.is_prev_push_key || self.stack.first) {
+								self.state = ReadProofState::Finished;
+								return Some(Err(VerifyError::ExtraneousNode)) // TODO a decode op error
+								              // return Err(CompactDecoderError::ValueNotAfterPush.
+								              // into())
+							}
+							self.stack.is_prev_push_key = false;
+							self.stack.is_prev_pop_key = false;
+							self.stack.is_prev_hash_child = None;
+							self.stack.first = false;
+						},
+						_ => {
+							self.stack.is_prev_push_key = false;
+							self.stack.is_prev_pop_key = false;
+							self.stack.is_prev_hash_child = None;
+							self.stack.first = false;
+						},
+					}
+
+					// debug TODO make it log and external function
+					match &op {
+						Op::HashChild(hash, child_ix) => {
+							println!(
+								"ChildHash {:?}, {:?}, {:?}",
+								self.stack.prefix, child_ix, hash
+							);
+						},
+						Op::HashValue(hash) => {
+							println!("ValueHash {:?}, {:?}", self.stack.prefix, hash);
+						},
+						Op::Value(value) => {
+							println!("Value {:?}, {:?}", self.stack.prefix, value);
+						},
+						_ => (),
+					}
+				}
+				from_iter = false;
+
+				// next
+				let item = if match &op {
+					Op::Value(..) | Op::HashValue(..) => true,
+					_ => false,
+				} {
+					let mut at_value = false;
+					let mut next_query = false;
+					if let Some(current) = self.current.as_ref() {
+						let query_slice = LeftNibbleSlice::new(&current.key);
+						match self.stack.prefix.as_leftnibbleslice().cmp(&query_slice) {
+							Ordering::Equal =>
+								if !self.stack.items.is_empty() {
+									at_value = true;
+								},
+							Ordering::Less => (),
+							Ordering::Greater =>
+								if current.as_prefix {
+									let query_slice = LeftNibbleSlice::new(&current.key);
+									if self
+										.stack
+										.prefix
+										.as_leftnibbleslice()
+										.starts_with(&query_slice)
+									{
+										at_value = true;
+									} else {
+										next_query = true;
+									}
+								} else {
+									next_query = true;
+								},
+						}
+						if next_query {
+							self.buf_op = Some(op);
+							self.state = ReadProofState::SwitchQueryPlan;
+							if current.as_prefix {
+								break
+							} else {
+								return Some(Ok(ReadProofItem::NoValue(&current.key)))
+							}
+						}
+					}
+
+					if at_value {
+						match &op {
+							Op::Value(value) => {
+								// TODO could get content from op with no clone.
+								Some(ReadProofItem::Value(
+									self.stack.prefix.inner().to_vec().into(),
+									value.clone(),
+								))
+							},
+							Op::HashValue(hash) => {
+								// TODO could get content from op with no clone.
+								Some(ReadProofItem::Hash(
+									self.stack.prefix.inner().to_vec().into(),
+									hash.clone(),
+								))
+							},
+							_ => unreachable!(),
+						}
+					} else {
+						match &op {
+							Op::Value(value) => {
+								// hash value here not value
+								if L::MAX_INLINE_VALUE
+									.map(|max| max as usize <= value.len())
+									.unwrap_or(false)
+								{
+									self.state = ReadProofState::Finished;
+									return Some(Err(VerifyError::ExtraneousValue(value.clone())))
+								}
+							},
+							_ => (),
+						}
+						None
+					}
+				} else {
+					None
+				};
+
+				// act
+				let r = match op {
+					Op::KeyPush(partial, mask) => {
+						self.stack
+							.prefix
+							.append_slice(LeftNibbleSlice::new_with_mask(partial.as_slice(), mask));
+						self.stack.stack_empty(self.stack.prefix.len());
+						Ok(())
+					},
+					Op::KeyPop(nb_nibble) => {
+						let r = self.stack.stack_pop(Some(nb_nibble as usize), &self.expected_root);
+						self.stack.prefix.drop_lasts(nb_nibble.into());
+						r
+					},
+					Op::EndProof => break,
+					op => self.stack.set_cache_change(op.into()),
+				};
+				if let Err(e) = r {
+					self.state = ReadProofState::Finished;
+					return Some(Err(e))
+				}
+				if let Some(r) = item {
+					if self.current.as_ref().map(|c| !c.as_prefix).unwrap_or(true) {
+						self.state = ReadProofState::SwitchQueryPlan; // TODO this is same as content NOne?
+					}
+					return Some(Ok(r))
+				}
+			}
+			if self.state != ReadProofState::SwitchQueryPlan && self.current.is_some() {
+				// TODO return halt instead
+				return Some(Err(VerifyError::IncompleteProof))
+			}
+			self.state = ReadProofState::SwitchQueryPlan;
+		}
+
+		/*
+		self.state = ReadProofState::Finished;
+		if self.proof.next().is_some() {
+			return Some(Err(VerifyError::ExtraneousNode))
+		} else {
+			return None
+		}
+			*/
+	}
+}
+
+/// When process is halted keep execution state
+/// to restore later.
+pub struct HaltedStateCheckContent<'a, L: TrieLayout, C> {
+	query_plan: QueryPlan<'a, C>,
+	current: Option<QueryPlanItem<'a>>,
+	stack: ReadContentStack<L>,
+	state: ReadProofState,
+}
+
+impl<'a, L: TrieLayout, C> From<QueryPlan<'a, C>> for HaltedStateCheckContent<'a, L, C> {
+	fn from(query_plan: QueryPlan<'a, C>) -> Self {
+		HaltedStateCheckContent {
+			stack: ReadContentStack {
+				items: Default::default(),
+				start_items: 0,
+				prefix: Default::default(),
+				expect_value: false,
+				iter_prefix: None,
+				is_prev_push_key: false,
+				is_prev_pop_key: false,
+				is_prev_hash_child: None,
+				first: true,
+				_ph: PhantomData,
 			},
-			Err(_) => Some(None),
+			state: ReadProofState::NotStarted,
+			current: None,
+			query_plan,
 		}
 	}
 }
